Adicionando a tela de lista de eventos e criando a navegacao da bottomnavigationbar
